cli: refactor help to the new approach
